Alias class moved to another package
